+++ Controller/Index/Request.php
@@ -26,21 +26,18 @@ class Request extends Index
      */
     public function execute()
     {
-        $this->writeLog();
+        $this->writeLog();        
+        $result   = new DataObject();
+        $response = $this->getResponse();
 
         if ($this->checkPost()) {
             $params = $this->generator->createRequest(
                 $this->getRequest()->getParam('paytype'),
                 $this->getRequest()->getParam('orderid')
             );
-
-            $result   = new DataObject();
-            $response = $this->getResponse();
             $result->addData($params);
-
-            return $response->representJson($result->toJson());
         }
 
-        die('No post!?');
+        return $response->representJson($result->toJson());
     }
 }